% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSDE}
\alias{TSDE}
\title{TSDE algorithm}
\usage{
TSDE(data, mean, sd, nmbcuts = 100L, lower00 = -3.5, upper00 = 3.5)
}
\arguments{
\item{data}{the input data (unstandardized)}

\item{nmbcuts}{number of possible splitting points}

\item{lower00}{lower bound of data (in units of standard deviations)}

\item{upper00}{upper bound of data (in units of standard deviations)}
}
\description{
TSDE algorithm
}
