% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSDE}
\alias{TSDE}
\title{TSDE algorithm}
\usage{
TSDE(
  data,
  nmbcuts = 100L,
  lower00 = -3.5,
  upper00 = 3.5,
  csmooth = 0.1,
  sizetre = 20L
)
}
\arguments{
\item{data}{the input data (unstandardized)}

\item{nmbcuts}{number of possible splitting points}

\item{lower00}{lower bound of data (in units of standard deviations)}

\item{upper00}{upper bound of data (in units of standard deviations)}

\item{csmooth}{the smoothing parameter: should be smaller than 0.7}

\item{sizetre}{the final tree size}
}
\description{
run the algorithm described in Spear, Grieb, Shang (1994); return a data frame object
}
\examples{
\dontrun{
library(tsde94)
data(sdata)
TSDE(sdata)
}

}
